requirements.txt
Appium-Python-Client~=2.11.1
miraelogger==0.0.2